refactor: Player 객체와 Name 객체 분리
