+++ src/main/java/Player.java
@@ -1,0 +1,11 @@
+public class Player {
+
+    Name name;
+
+    public Player(Name name) {
+        this.name = name;
+    }
+
+
+
+}
